worker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-firestore
google-cloud-storage
google-cloud-vision==3.4.5
python-multipart==0.0.6
jinja2==3.1.3
python-docx==1.1.2
